+++ workalendar/asia/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from .china import China
 from .hong_kong import HongKong
-from .japan import Japan
+from .japan import Japan, JapanBank
 from .malaysia import Malaysia
 from .qatar import Qatar
 from .singapore import Singapore
@@ -14,6 +14,7 @@ __all__ = (
     'China',
     'HongKong',
     'Japan',
+    'JapanBank',
     'Malaysia',
     'Qatar',
     'Singapore',