+++ Changelog.md
@@ -2,7 +2,11 @@
 
 ## master (unreleased)
 
-Nothing to see here yet.
+- Added adjustments to 2019-2020 Japan calendar due to the coronation of a new emperor.
+
+### New Calendar
+
+- Added JapanBank by @raybuhr (#379)
 
 ## v5.1.1 (2019-06-27)